remove type pipe
